+++ packages/generator/scripts/xsanity/handlers/common.ts
@@ -50,3 +50,8 @@ export function ObjMure3(checks: Check[], ra: RoomActor) {
 export function ObjHamishi(checks: Check[], ra: RoomActor) {
   checks.push({ roomActor: ra, item: 'NOTHING', name: 'Red Boulder', type: 'boulder-red' });
 }
+
+export function EnGs(checks: Check[], ra: RoomActor) {
+  checks.push({ roomActor: ra, item: 'FAIRY', name: 'Gossip Fairy', type: 'gossip', sliceId: 0 });
+  checks.push({ roomActor: ra, item: 'FAIRY_BIG', name: 'Gossip Big Fairy', type: 'gossip-big', sliceId: 1 });
+}